Break apart test files
